refactor(racinggame): Car 클래스 분리
