+++ src/main/java/racingcar/model/Car.java
@@ -1,0 +1,24 @@
+package racingcar.model;
+
+public class Car {
+    private String name;
+    private int position = 0;
+
+    public Car(String name) {
+        this.name = name;
+    }
+
+    public void move() {
+        if (Math.random() > 0.4) {
+            position++;
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getPosition() {
+        return position;
+    }
+}
